cachedb_dynamodb: Fix ARM32 compile warnings

+++ modules/cachedb_dynamodb/cachedb_dynamodb_dbase.c
@@ -304,7 +304,7 @@ int dynamodb_map_set(cachedb_con *connection, const str *key, const str *keyset,
 				LM_ERR("No more pkg mem\n");
 				return -1;
 			}
-			sprintf(attribute_value_int, "%ld", pair->val.val.i64);
+			sprintf(attribute_value_int, "%lld", (long long)pair->val.val.i64);
 			init_str(&attribute_value, attribute_value_int);
 			break;
